+++ BizHawk.Emulation.Cores/Computers/AmstradCPC/AmstradCPC.Messaging.cs
@@ -430,7 +430,7 @@ namespace BizHawk.Emulation.Cores.Computers.AmstradCPC
 			int end = _machine.TapeDevice.DataBlocks[_machine.TapeDevice.CurrentDataBlockIndex].DataPeriods.Count;
 			double p = 0;
 			if (end != 0)
-				p = ((double)pos / (double)end) * (double)100;
+				p = ((double)pos / (double)end) * 100.0;
 
 			sb.Append(p.ToString("N0") + "%");
 			SendMessage(sb.ToString().TrimEnd('\n'), MessageCategory.Tape);
@@ -454,7 +454,7 @@ namespace BizHawk.Emulation.Cores.Computers.AmstradCPC
 			}
 			// work out overall position within the tape
 			p = 0;
-			p = ((double)ourPos / (double)cnt) * (double)100;
+			p = ((double)ourPos / (double)cnt) * 100.0;
 			sb.Append(p.ToString("N0") + "%");
 			SendMessage(sb.ToString().TrimEnd('\n'), MessageCategory.Tape);
 		}
